rename: shared 폴더의 icons -> icon
