+++ client/src/shared/icon/InearLogo.tsx
