+++ client/src/shared/icon/SendIcon.tsx
