+++ client/src/widgets/vote/Vote.tsx
@@ -1,4 +1,4 @@
-import { ChevronDown } from '@/shared/icons/ChevronDown';
+import { ChevronDown } from '@/shared/icon/ChevronDown';
 import { useState } from 'react';
 
 export function Vote() {
